+++ docs/demo/table/editable-events.vue
@@ -1,7 +1,7 @@
 <!-- eslint-disable no-console -->
 <script lang="ts" setup>
 import { ref } from 'vue'
-import type { EditableTableEventParams } from '@ideaz/element'
+import type { EditableTableEventParams } from 'ideaz-element'
 
 interface RowData {
   name: string