+++ index.ts
@@ -1,4 +1,4 @@
-export * from '@ideaz/element'
+export * from 'ideaz-element'
 export * from '@ideaz/hooks'
 export * from '@ideaz/directives'
 export * from '@ideaz/shared'